task: moved stats functions to stats folder
