--- Calculators/PopulationMean.py
@@ -1,2 +1,0 @@
-def population_mean(data):
-    return float(sum(data)) / max(len(data), 1)
