+++ Statistics/PopulationStandardDeviation.py
@@ -1,4 +1,4 @@
-from Calculators.PopulationMean import population_mean
+from Statistics.PopulationMean import population_mean
 from Calculators.SquareRoot import square_root
 
 
@@ -6,8 +6,6 @@ def pop_stand_dev(data):
     n = len(data)
     u = population_mean(data)
     return square_root(sum([(element - u) ** 2 for element in data]) / (len(data) - 1))
-
-
