+++ Statistics/SampleStandardDeviation.py
@@ -1,5 +1,6 @@
 import math
 
+
 def sampleSTD(s, i, x, n):
     s = math.sqrt(sum(i - x) * (i - x) / (n - 1))
     return s
