+++ Statistics/Statistics.py
@@ -1,14 +1,36 @@
 from Calculators.Calculator import Calculator
-from Calculators.ZScore import z_score
-from Calculators.ConfidenceInterval import confidence_interval
-from Calculators.PopulationVar import population_variance
+from Statistics.ZScore import z_score
+from Statistics.ConfidenceInterval import confidence_interval
+from Statistics.PopulationVar import population_variance
+from Statistics.PopulationMean import population_mean
+from Statistics.Median import median
+from Statistics.Mode import mode
+from Statistics.VariancePopulationProportion import var_pop_proportion
+from CSVReader.CSVReader import CsvReader
 
 
 class Statistics(Calculator):
     data = []
 
     def __init__(self):
-        pass
+        self.data = CsvReader(filepath)
+        super().__init__()
+
+    def pop_mean(self):
+        self.result = population_mean(self.data)
+        return self.result
+
+    def med(self):
+        self.result = median(self.data)
+        return self.result
+
+    def mod(self):
+        self.result = mode(self.data)
+        return self.result
+
+    def variance_pop_proportion(self):
+        self.result = var_pop_proportion(self.data)
+        return self.result
 
     def z_score(self):
         self.result = z_score()
@@ -21,4 +43,3 @@ class Statistics(Calculator):
     def population_variance(self, d):
         self.result = population_variance(d)
         return self.result
-
